New option for convert-menu font->curve: "Curve (Single filling group)".
Since the text-object remake, it was not possible anymore to 'cut out'
text from, say, a box made by a polycurve, by converting text to curve
and then joining - only the first character would be cut out.

This is because of the filling groups (nu/dl->charidx)
I introduced for getting the vast speedup and the possibility
of overlapping characters.

The new convert menu option now assigns filling group 0 to all of
the nurbs generated.

Maybe filling groups should be exposed in the UI in general for curves -
there are various occasions where they are useful.

(Hint to the UI mafia! ;)