Armature Drawing Tweak:
I've always found it hard to see the armature bones in 'Line' mode that don't have any constraints/ik/keyframes/etc. colour indications as they were set to the colour of the background. This tweak makes these particular bones draw slightly darker than
the minor-gridlines in the 3d-view, but not as dark as the major gridlines.