Quite a few bugfixes:
* Constraint sliders in Action Editor didn't update buttons window

* Homekey in IPO Editor didn't respect scaling.

* With Preview Range set, using the SKEY and EKEY hotkeys in the Timeline
to change the extents of it, resulted to the NLA editor not being
refreshed. (Note: I haven't actually seen this happening)

* Keyframing options 'Avail' and 'Needed' didn't work for Objects
when they had an ObAction. The problem was that the code assumed
that the object had an ipo-block associated with it, but that is
not the case when you have an Object Action for that object.