== Clamp To Constraint ==
Now when editing the curve that an Object with a Clamp To Constraint uses as its target, the Object's position will be updated in realtime to match to changes in the curve. 

I was going to have this working for bones too, but I haven't got that quite working yet.