This is patch [#6772] Background image drawing speedup Thanks goto Matthew Plough (meestaplu)
I've tested it on linux with both an nvidia and an ATI card,
I've also tested it on a mac and solaris, as well as through a
remote display.  So I think this one is pretty safe.  
If it causes problems on older systems we can revert it
but I don't think we will have any issues with it.

Kent